Move new features to qol.cpp
