+++ Source/all.h
@@ -20,6 +20,8 @@
 
 #include "../types.h"
 
+#include "../SourceX/qol.h"
+
 #include "appfat.h"
 #include "automap.h"
 #include "capture.h"
