+++ Source/automap.h
@@ -35,15 +35,6 @@ void AutomapRight();
 void AutomapZoomIn();
 void AutomapZoomOut();
 void DrawAutomap();
-void DrawMonsterHealthBar(int monsterID);
-void DrawXPBar();
-void HighlightItemsNameOnMap();
-void AddItemToDrawQueue(int x, int y, int id);
-void diablo_parse_config();
-extern bool altPressed;
-extern bool drawXPBar;
-extern bool drawHPBar;
-extern int highlightItemsMode;
 void DrawAutomapTile(int screen_x, int screen_y, WORD automap_type);
 #ifdef HELLFIRE
 void SearchAutomapItem();
